// src/app/models/interfaces/apigee-token.ts
export interface ApigeeToken {
  issued_at: string;
  scope: string;
  application_name: string;
  refresh_token_issued_at: string;
  status: string;
  refresh_token_status: string;
  api_product_list: string;
  expires_in: string;
  'developer.email': string;
  organization_id: string;
  token_type: string;
  refresh_token: string;
  client_id: string;
  access_token: string;
  organization_name: string;
  refresh_token_expires_in: string;
  refresh_count: string;
}